:truck: Creating new sidebar for Flights Report
