+++ src/app/components/sidebar/sidebar.component.ts
@@ -9,7 +9,7 @@ declare interface RouteInfo {
 }
 export const ROUTES: RouteInfo[] = [
     { path: '/dashboard', title: 'Dashboard',  icon: 'dashboard', class: '' },
-    // { path: '/user-profile', title: 'Ingresar Vuelo',  icon:'note_add', class: '' },
+    { path: '/user-profile', title: 'Flight Report',  icon:'analytics', class: '' },
     // { path: '/table-list', title: 'Monitorizar',  icon:'content_paste', class: '' },
     //{ path: '/typography', title: 'Typography',  icon:'library_books', class: '' },
     //{ path: '/icons', title: 'Icons',  icon:'bubble_chart', class: '' },
